requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
astunparse==1.6.3
attrs==24.2.0
bleak==0.22.3
cachetools==5.5.0
cattrs==24.1.2
certifi==2024.8.30
charset-normalizer==3.4.0
chex==0.1.7
contourpy==1.3.1
coremltools==8.1
cycler==0.12.1
dbus-fast==2.24.4
dm-tree==0.1.8
etils==1.11.0
filelock==3.16.1
flatbuffers==24.3.25
flax==0.10.2
fonttools==4.55.2
fsspec==2024.10.0
gast==0.6.0
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-generativeai==0.8.3
google-pasta==0.2.0
googleapis-common-protos==1.66.0
grpcio==1.68.0
grpcio-status==1.68.0
h5py==3.12.1
httplib2==0.22.0
humanize==4.11.0
idna==3.10
importlib_resources==6.4.5
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.4
joblib==1.4.2
keras==2.15.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.3
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
msgpack==1.1.0
ncnn==1.0.20240820
nest_asyncio==1.6.0
networkx==3.4.2
numpy==1.23.5
oauthlib==3.2.2
onnx==1.17.0
opencv-python==4.10.0.84
openvino==2024.5.0
openvino-telemetry==2024.5.0
opt_einsum==3.4.0
optax==0.2.1
orbax-checkpoint==0.6.4
packaging==23.2
paho-mqtt==2.1.0
pandas==2.2.3
pillow==11.0.0
portalocker==3.0.0
proto-plus==1.25.0
protobuf==5.29.0
psutil==6.1.0
py-cpuinfo==9.0.0
pyaml==24.9.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.2
pydantic_core==2.27.1
Pygments==2.18.0
pyparsing==3.2.0
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
six==1.17.0
sympy==1.13.1
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.1
tensorflow-cpu-aws==2.15.1
tensorflow-estimator==2.15.0
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.37.1
tensorflow_decision_forests==1.8.1
tensorflowjs==4.22.0
tensorstore==0.1.69
termcolor==2.5.0
tf_keras==2.15.1
threadpoolctl==3.5.0
toolz==1.0.0
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.44
ultralytics-thop==2.0.12
uritemplate==4.1.1
urllib3==2.2.3
Werkzeug==3.1.3
wrapt==1.14.1
wurlitzer==3.1.1
zipp==3.21.0
